Fix second interface networking, again
Despite the previous fixes, after a recent update the ethernet
member of the bridge interface was getting both an ip and a ipv6
address. This appears to be due to change in the behaviour of
libvirt and/or accept_ra upstream (most likely due to some security
issue).

This commit disables ipv6 on the bridge and on the member ethernet
interface, in addition to avoiding either getting an ip address.
This results in the desired bridge-only behaviour for the bridge
on the second interface.

Signed-off-by: Daniel F. Dickinson <dfdpublic@wildtechgarden.ca>